Create Const & Utils directories
