+++ src/components/board.js
@@ -1,4 +1,4 @@
-import {createElement} from "../helpers";
+import createElement from "../utils/createElement";
 
 const createBoardTemplate = () => {
   return (
