+++ src/components/menu.js
@@ -1,4 +1,4 @@
-import {createElement} from "../helpers";
+import createElement from "../utils/createElement";
 
 const createMenuItemTemplate = (item) => {
   const {id, title, isChecked} = item;
