+++ src/const/colors.js
@@ -1,0 +1,7 @@
+export default [
+  `black`,
+  `yellow`,
+  `blue`,
+  `green`,
+  `pink`
+];
