+++ src/const/days.js
@@ -1,0 +1,9 @@
+export default [
+  `mo`,
+  `tu`,
+  `we`,
+  `th`,
+  `fr`,
+  `sa`,
+  `su`
+];
