+++ src/const/render-position.js
@@ -1,0 +1,4 @@
+export default {
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`
+};
