+++ src/utils/createElement.js
@@ -1,0 +1,6 @@
+export default (template) => {
+  const newElement = document.createElement(`div`);
+  newElement.innerHTML = template;
+
+  return newElement.firstElementChild;
+};
