+++ src/utils/getRandomInt.js
@@ -1,0 +1,1 @@
+export default (min, max) => min + Math.floor(Math.random() * (max + 1 - min));
