+++ src/utils/render.js
@@ -1,0 +1,12 @@
+import {RenderPosition} from "../const/render-position";
+
+export default (container, element, place = RenderPosition.BEFOREEND) => {
+  switch (place) {
+    case RenderPosition.AFTERBEGIN:
+      container.prepend(element);
+      break;
+    case RenderPosition.BEFOREEND:
+      container.append(element);
+      break;
+  }
+};
